# Sdf scaling