--- a/assets/shaders/gi_ss_blend.wgsl
+++ b/assets/shaders/gi_ss_blend.wgsl
@@ -7,18 +7,19 @@
 @group(0) @binding(0) var<uniform> camera_params:     CameraParams;
 @group(0) @binding(1) var<uniform> state:             GiState;
 @group(0) @binding(2) var<storage> probes:            ProbeDataBuffer;
-@group(0) @binding(3) var          sdf_in:            texture_storage_2d<r16float,    read>;
-@group(0) @binding(4) var          ss_bounce_in:      texture_storage_2d<rgba32float, read>;
-@group(0) @binding(5) var          ss_blend_out:      texture_storage_2d<rgba32float, write>;
+@group(0) @binding(3) var          sdf_in:            texture_2d<f32>;
+@group(0) @binding(4) var          sdf_in_sampler:    sampler;
+@group(0) @binding(5) var          ss_bounce_in:      texture_storage_2d<rgba32float, read>;
+@group(0) @binding(6) var          ss_blend_out:      texture_storage_2d<rgba32float, write>;
 
 fn distance_squared(a: vec2<f32>, b: vec2<f32>) -> f32 {
     let c = a - b;
     return dot(c, c);
 }
 
-fn get_sdf_screen(screen_pose: vec2<i32>) -> f32 {
-    return textureLoad(sdf_in, screen_pose).r;
-}
+// fn get_sdf_screen(screen_pose: vec2<i32>) -> f32 {
+//     return textureLoad(sdf_in, screen_pose, 0).r;
+// }
 
 fn raymarch_occlusion(
     ray_origin:    vec2<f32>,
@@ -37,7 +38,9 @@ fn raymarch_occlusion(
         }
 
         let h          = ray_origin + ray_progress * ray_direction;
-        let scene_dist = get_sdf_screen(world_to_screen(h, camera_params.screen_size, camera_params.view_proj));
+        
+        let uv = world_to_sdf_uv(h, camera_params.view_proj, camera_params.inv_sdf_scale);
+        let scene_dist = bilinearSample(0, sdf_in, sdf_in_sampler, uv);
 
         if (scene_dist <= 0.1) {
             return 0.0;