--- a/src/gi/gi_component.rs
+++ b/src/gi/gi_component.rs
@@ -1,15 +1,15 @@
-use bevy::prelude::{Component, Vec2, Color};
-use bevy_inspector_egui::Inspectable;
 use bevy::prelude::*;
+use bevy::prelude::{Color, Component, Vec2};
+use bevy_inspector_egui::Inspectable;
 
 #[derive(Reflect, Component, Clone, Copy, Default, Inspectable)]
 #[reflect(Component)]
 pub struct LightSource {
-    pub radius:    f32,
+    pub radius: f32,
     pub intensity: f32,
-    pub color:     Color,
-    pub falloff:   Vec3,
-    pub jitter_intensity:   f32,
+    pub color: Color,
+    pub falloff: Vec3,
+    pub jitter_intensity: f32,
     pub jitter_translation: f32,
 }
 
@@ -32,6 +32,6 @@ pub struct AmbientMask {
 #[derive(Reflect, Component, Clone, Copy, Default, Inspectable)]
 #[reflect(Component)]
 pub struct GiAmbientLight {
-    pub color:     Color,
+    pub color: Color,
     pub intensity: f32,
 }