--- a/src/gi/gi_gpu_types.rs
+++ b/src/gi/gi_gpu_types.rs
@@ -1,28 +1,28 @@
-use crate::SCREEN_SIZE;
 use crate::gi::gi_component::LightSource;
+use crate::SCREEN_SIZE;
 use bevy::prelude::{Mat4, Vec2, Vec3};
 use bevy::render::render_resource::ShaderType;
 
-use super::gi_config::{GI_SDF_JITTER_CONTRIB, GI_SDF_MAX_STEPS, GI_SCREEN_PROBE_SIZE};
+use super::gi_config::{GI_SCREEN_PROBE_SIZE, GI_SDF_JITTER_CONTRIB, GI_SDF_MAX_STEPS};
 
 #[derive(Default, Clone, ShaderType)]
 pub(crate) struct GiGpuLightSource {
-    pub(crate) center:    Vec2,
+    pub(crate) center: Vec2,
     pub(crate) intensity: f32,
-    pub(crate) color:     Vec3,
-    pub(crate) radius:    f32,
-    pub(crate) falloff:   Vec3,
+    pub(crate) color: Vec3,
+    pub(crate) radius: f32,
+    pub(crate) falloff: Vec3,
 }
 
 impl GiGpuLightSource {
     pub fn new(light: LightSource, center: Vec2) -> Self {
         let color = light.color.as_rgba_f32();
         Self {
-            center:    center,
-            radius:    light.radius,
+            center: center,
+            radius: light.radius,
             intensity: light.intensity,
-            color:     Vec3::new(color[0], color[1], color[2]),
-            falloff:   light.falloff,
+            color: Vec3::new(color[0], color[1], color[2]),
+            falloff: light.falloff,
         }
     }
 }
@@ -55,21 +55,21 @@ pub(crate) struct GiGpuLightOccluderBuffer {
 
 #[derive(Default, Clone, ShaderType)]
 pub(crate) struct GiGpuCameraParams {
-    pub(crate) screen_size:       Vec2,
-    pub(crate) screen_size_inv:   Vec2,
-    pub(crate) view_proj:         Mat4,
+    pub(crate) screen_size: Vec2,
+    pub(crate) screen_size_inv: Vec2,
+    pub(crate) view_proj: Mat4,
     pub(crate) inverse_view_proj: Mat4,
 }
 
 #[derive(Clone, ShaderType, Debug)]
 pub(crate) struct GiGpuState {
-    pub gi_frame_counter:   i32,
-    pub ss_probe_size:      i32,
-    pub ss_atlas_cols:      i32,
-    pub ss_atlas_rows:      i32,
-    pub sdf_max_steps:      i32,
+    pub gi_frame_counter: i32,
+    pub ss_probe_size: i32,
+    pub ss_atlas_cols: i32,
+    pub ss_atlas_rows: i32,
+    pub sdf_max_steps: i32,
     pub sdf_jitter_contrib: f32,
-    pub gi_ambient:         Vec3,
+    pub gi_ambient: Vec3,
 }
 
 impl Default for GiGpuState {
@@ -103,10 +103,15 @@ impl Default for GiGpuProbeDataBuffer {
         let cols = SCREEN_SIZE.0 / (GI_SCREEN_PROBE_SIZE as usize);
         let rows = SCREEN_SIZE.1 / (GI_SCREEN_PROBE_SIZE as usize);
         let size = cols * rows;
-        return Self{
+        return Self {
             count: size as u32,
-            data:  vec![GiGpuProbeData { camera_pose: Vec2::ZERO }; size],
-        }
+            data: vec![
+                GiGpuProbeData {
+                    camera_pose: Vec2::ZERO
+                };
+                size
+            ],
+        };
     }
 }