This change adds the ability to scale the area the sdf texture represents and then uses bilinear interpolation in the shaders that read the sdf data. This gives an easily adjustable way to address offscreen shadow issues

Before Change
![image](https://user-images.githubusercontent.com/9063043/210025656-7a4f7041-b7d2-444b-b7e3-52457a9b1dec.png)

After Change
![image](https://user-images.githubusercontent.com/9063043/210025607-be299118-51a4-415b-b1e8-b779e4c0171f.png)