@dbenson24 2022-12-30
I also am seeing a performance boost with this change.
Before
```
2022-12-30T01:33:50.445300Z  INFO bevy diagnostic: frame_count                     : 689.000000
2022-12-30T01:33:51.444401Z  INFO bevy diagnostic: frame_time                      :   19.929964ms (avg 20.097590ms)
2022-12-30T01:33:51.444733Z  INFO bevy diagnostic: fps                             :   50.280973   (avg 49.885925)
```

After
```
2022-12-30T01:32:25.658418Z  INFO bevy diagnostic: frame_count                     : 1408.000000
2022-12-30T01:32:26.653097Z  INFO bevy diagnostic: frame_time                      :   18.884080ms (avg 18.671375ms)
2022-12-30T01:32:26.653402Z  INFO bevy diagnostic: fps                             :   53.029591   (avg 53.649381)
```